Extracted createTokenIterator
