+++ src/main/index.ts
@@ -10,4 +10,5 @@ export {seq} from './seq';
 export {text, TextOptions} from './text';
 export {until, UntilOptions} from './until';
 export * from './taker-types';
+export * from './token-utils';
 export * from './Tokenizer';
